Fix for API not responding...
The file parser cut off the last letter of the uuid and that caused the api to not recognise the uuid and the program used to copy the data from the wrong player due to a list missuse. those have been fixed. the api link also changed due to unreliabilty from the past one